Implement a new option ("-hub") to work aggressively to acquire connections
When you're mining, it's vital that you be informed of new blocks on the
network as quickly as possible. This ensures that you are working on what is
most likely to be the longest chain. Similarly, it is vital that you get
your own blocks out to the network as soon as possible. This makes it the
most likely that your blocks will be incorporated in the chain that
ultimately wins.

Yet the code in bitcoind is stunningly conservative. It prioritizes
minimizing connections and overhead in a way that's great for a personal
wallet that does transactions but horrible for a mining controller. It is
also very slow to recover connections on a restart.

A new option "-hub" is added. It takes a mode option that can be 1 (small
hub), 2 (medium hub), or 3 (large hub). As you increase the hub level, the
code will work more aggressively to acquire more connections. Note that
resource consumption will be increased.

In addition, the code that determines when to process the list of connected
nodes for received messages has been made smarter.