Fix off-by-one errors in using IsFinal()
Previous AcceptBlock() and CreateNewBlock() didn't take into account the
fact that IsFinal() without any arguments tests if the transaction is
considered final in the *current* block, when both those functions
really needed to know if the transaction would be final in the *next*
block.

Additionally the UI had a similar misunderstanding.